+++ CHANGELOG.md
@@ -7,6 +7,8 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 ## [unreleased]
 ### Changed
 - Archive prod docker image with release tag name. Update action versions.
+### Fixed
+- Error image background static path
 
 ## [2.2]
 ### Added