+++ CHANGELOG.md
@@ -7,6 +7,8 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 ## [unreleased]
 ### Added
 - Link out to Scout: introduce config variable for base URL
+- Link out to Scout: case links on home sample list
+- 
 ### Changed
 - Archive prod docker image with release tag name. Update action versions.
 ### Fixed