+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/)
 
 About changelog [here](https://keepachangelog.com/en/1.0.0/)
 
-## [unreleased]
+## [2.3]
 ### Added
 - Link out to Scout: introduce config variable for base URL
 - Link out to Scout: case links on home sample list